+++ src/programs/raster/mod.rs
@@ -1,2 +1,9 @@
+#[cfg(all(major_ge_3, minor_ge_1))]
+mod mdimtranslate;
 mod vrt;
+
+#[cfg(all(major_ge_3, minor_ge_1))]
+pub use mdimtranslate::{
+    multi_dim_translate, MultiDimTranslateDestination, MultiDimTranslateOptions,
+};
 pub use vrt::*;